Using int16 for left and right drive members. This is needed to have forward/reverse with 0-255 in each direction.

+++ Command.hpp
@@ -23,8 +23,8 @@ struct Command
         struct DriveCmd 
         {
             uint8_t type;
-            uint8_t left;
-            uint8_t right;
+            int16_t left;
+            int16_t right;
         } drive;
     };
